impl FstOp, FstOp2, FstCache for Deref instead of Arc to be more generic
