+++ rustfst/src/algorithms/lazy/cache/fst_cache.rs
@@ -1,4 +1,5 @@
 use std::fmt::Debug;
+use std::ops::Deref;
 
 use crate::algorithms::lazy::CacheStatus;
 use crate::semirings::Semiring;
@@ -22,4 +23,54 @@ pub trait FstCache<W: Semiring>: Debug {
 
     fn len_trs(&self) -> usize;
     fn len_final_weights(&self) -> usize;
+}
+
+impl<W: Semiring, C: FstCache<W>, CP: Deref<Target=C> + Debug> FstCache<W> for CP {
+    fn get_start(&self) -> CacheStatus<Option<usize>> {
+        self.deref().get_start()
+    }
+
+    fn insert_start(&self, id: Option<usize>) {
+        self.deref().insert_start(id)
+    }
+
+    fn get_trs(&self, id: usize) -> CacheStatus<TrsVec<W>> {
+        self.deref().get_trs(id)
+    }
+
+    fn insert_trs(&self, id: usize, trs: TrsVec<W>) {
+        self.deref().insert_trs(id, trs)
+    }
+
+    fn get_final_weight(&self, id: usize) -> CacheStatus<Option<W>> {
+        self.deref().get_final_weight(id)
+    }
+
+    fn insert_final_weight(&self, id: usize, weight: Option<W>) {
+        self.deref().insert_final_weight(id, weight)
+    }
+
+    fn num_known_states(&self) -> usize {
+        self.deref().num_known_states()
+    }
+
+    fn num_trs(&self, id: usize) -> Option<usize> {
+        self.deref().num_trs(id)
+    }
+
+    fn num_input_epsilons(&self, id: usize) -> Option<usize> {
+        self.deref().num_input_epsilons(id)
+    }
+
+    fn num_output_epsilons(&self, id: usize) -> Option<usize> {
+        self.deref().num_output_epsilons(id)
+    }
+
+    fn len_trs(&self) -> usize {
+        self.deref().len_trs()
+    }
+
+    fn len_final_weights(&self) -> usize {
+        self.deref().len_final_weights()
+    }
 }
