+++ rustfst/src/algorithms/lazy/fst_op.rs
@@ -1,7 +1,5 @@
 use std::fmt::Debug;
 use std::ops::Deref;
-use std::rc::Rc;
-use std::sync::Arc;
 
 use anyhow::Result;
 
@@ -19,26 +17,7 @@ pub trait FstOp<W: Semiring>: Debug {
     fn properties(&self) -> FstProperties;
 }
 
-impl<W: Semiring, F: FstOp<W>> FstOp<W> for Arc<F> {
-    fn compute_start(&self) -> Result<Option<StateId>> {
-        self.deref().compute_start()
-    }
-
-    fn compute_trs(&self, id: usize) -> Result<TrsVec<W>> {
-        self.deref().compute_trs(id)
-    }
-
-    fn compute_final_weight(&self, id: StateId) -> Result<Option<W>> {
-        self.deref().compute_final_weight(id)
-    }
-
-    // Computed at construction time
-    fn properties(&self) -> FstProperties {
-        self.deref().properties()
-    }
-}
-
-impl<W: Semiring, F: FstOp<W>> FstOp<W> for Rc<F> {
+impl<W: Semiring, F: FstOp<W>, FP: Deref<Target=F> + Debug> FstOp<W> for FP {
     fn compute_start(&self) -> Result<Option<StateId>> {
         self.deref().compute_start()
     }
